feat: [CI-4765]: circular dependency CIStep
