+++ src/modules/75-ci/components/PipelineSteps/CIStep/StepUtils.tsx
@@ -6,10 +6,15 @@
  */
 
 import React from 'react'
-import { FormInput, MultiTypeInputType, Container, Text } from '@wings-software/uicore'
-import { FontVariation } from '@harness/design-system'
-import cx from 'classnames'
+import type { FormikErrors } from 'formik'
 import { get, isEmpty, set } from 'lodash-es'
+import cx from 'classnames'
+import { FontVariation } from '@harness/design-system'
+import { FormInput, MultiTypeInputType, Container, Text } from '@wings-software/uicore'
+import type { UseFromStageInfraYaml } from 'services/ci'
+import type { StringsMap } from 'stringTypes'
+import { usePipelineContext } from '@pipeline/components/PipelineStudio/PipelineContext/PipelineContext'
+import type { BuildStageElementConfig, StageElementWrapper } from '@pipeline/utils/pipelineTypes'
 import type { ConnectorInfoDTO } from 'services/cd-ng'
 import type { UseStringsReturn } from 'framework/strings'
 import { isRuntimeInput, CodebaseTypes } from '@pipeline/utils/CIUtils'
@@ -20,6 +25,62 @@ import {
 } from '@pipeline/components/PipelineInputSetForm/CICodebaseInputSetForm'
 import { FormMultiTypeRadioGroupField } from '@common/components/MultiTypeRadioGroup/MultiTypeRadioGroup'
 import css from '@pipeline/components/PipelineSteps/Steps/Steps.module.scss'
+export const useGetPropagatedStageById = (
+  stageId: string
+): StageElementWrapper<BuildStageElementConfig> | undefined => {
+  const { getStageFromPipeline } = usePipelineContext()
+
+  const { stage: currentStage } = getStageFromPipeline<BuildStageElementConfig>(stageId || '')
+
+  const isPropagatedStage = !isEmpty((currentStage?.stage?.spec?.infrastructure as UseFromStageInfraYaml)?.useFromStage)
+  const { stage: propagatedStage } = getStageFromPipeline<BuildStageElementConfig>(
+    (currentStage?.stage?.spec?.infrastructure as UseFromStageInfraYaml)?.useFromStage
+  )
+  return isPropagatedStage ? propagatedStage : currentStage
+}
+
+export const validateConnectorRefAndImageDepdendency = (
+  connectorRef: string,
+  image: string,
+  getString: (key: keyof StringsMap, vars?: Record<string, any> | undefined) => string
+): FormikErrors<any> => {
+  const errors: FormikErrors<any> = {}
+  if (connectorRef && !image) {
+    errors['spec.image'] = getString('ci.buildInfra.awsVM.isRequiredWhen', {
+      field1: getString('imageLabel'),
+      field2: getString('pipelineSteps.connectorLabel')
+    })
+  } else if (!connectorRef && image) {
+    errors['spec.connectorRef'] = getString('ci.buildInfra.awsVM.isRequiredWhen', {
+      field2: getString('imageLabel'),
+      field1: getString('pipelineSteps.connectorLabel')
+    })
+  }
+  return errors
+}
+
+export const AllMultiTypeInputTypesForStep = [
+  MultiTypeInputType.FIXED,
+  MultiTypeInputType.EXPRESSION,
+  MultiTypeInputType.RUNTIME
+]
+
+export const AllMultiTypeInputTypesForInputSet = [MultiTypeInputType.FIXED, MultiTypeInputType.EXPRESSION]
+
+/* Field of type lists have some limitations to support all three input types */
+
+/* a field of type list cannot assume expression as supported a value */
+export const SupportedInputTypesForListTypeField = [MultiTypeInputType.FIXED, MultiTypeInputType.RUNTIME]
+
+/* Note:  list items do not support runtime inputs by design, captured in https://harness.atlassian.net/browse/PIE-2617 */
+
+/* for few fields, list items cannot support be expressions due to a limitation, captured in https://harness.atlassian.net/browse/CI-3950 */
+export const SupportedInputTypesForOPVarsListItems = [MultiTypeInputType.FIXED]
+
+/* few fields are able to support expressions for list items */
+export const SupportedInputTypesForListItems = [MultiTypeInputType.FIXED, MultiTypeInputType.EXPRESSION]
+
+export const SupportedInputTypesForListTypeFieldInInputSetView = [MultiTypeInputType.FIXED]
 
 export const renderBuildTypeInputField = ({
   type,
